// Client/src/pages/DetailPost/types.ts
export interface PostDetailType {
  attractionAddress: string;
  attractionId: number;
  attractionName: string;
  memberId: number;
  commentCount: number;
  createdAt: string;
  isVoted: boolean;
  likes: number;
  modifiedAt: string;
  picture: string;
  postContents: string[];
  postHashTags: string[];
  postId: number;
  postImageUrls: string[];
  postTitle: string;
  username: string;
  views: number;
}